requirements.txt
# 웹 크롤링 및 파싱
requests==2.31.0
beautifulsoup4==4.13.4
lxml==5.1.0

# 데이터베이스
psycopg2-binary==2.9.10

# 비동기 처리 (선택사항)
aiohttp==3.12.15
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
async-timeout==5.0.1

# 스케줄링 (선택사항)
APScheduler==3.11.0
tzlocal==5.3.1

# RSS 피드 처리 (선택사항)
feedparser==6.0.11

# 재시도 로직 (선택사항)
tenacity==9.1.2

# 유틸리티
attrs==25.3.0
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
propcache==0.3.2
sgmllib3k==1.0.0
soupsieve==2.7
typing_extensions==4.14.1
yarl==1.20.1

# 개발 도구 (선택사항)
pytest==8.0.0
black==24.1.1
flake8==7.0.0
